add new session api over websockets, add dummy local session class, add websocket mock dev server, add message types as typescript doc

+++ lib/models/api/api_route.dart
@@ -22,10 +22,10 @@ class ApiRoute {
 
   ApiRoute.fromJson(Map<String, dynamic> json) {
     status = json['status'];
-    distance = json['distance'];
+    distance = json['distance'].toDouble();
     time = json['time'];
-    ascend = json['ascend'];
-    descend = json['descend'];
+    ascend = json['ascend'].toDouble();
+    descend = json['descend'].toDouble();
     if (json['points'] != null) {
       points = new List<ApiPoint>();
       json['points'].forEach((v) {
